Proportion function should return double not int

+++ source/misc.h
@@ -15,4 +15,4 @@ std::string getPath(std::string filename);
 
 unsigned long factorial(unsigned long n);
 unsigned int choose(unsigned long n, unsigned long choose);
-int proportion(int n);
+double proportion(int n);
